Correcciones

+++ src/database/seeds/TruncateSeeder.php
@@ -28,22 +28,22 @@ class TruncateSeeder extends Seeder {
         \Solunes\Master\App\Activity::truncate();
         \Solunes\Master\App\EmailTranslation::truncate();
         \Solunes\Master\App\Email::truncate();
-        \Solunes\Master\App\AlertUser::truncate();
-        \Solunes\Master\App\AlertConditional::truncate();
-        \Solunes\Master\App\AlertAction::truncate();
-        \Solunes\Master\App\Alert::truncate();
+        if(config('solunes.alerts')){
+            \Solunes\Master\App\AlertUser::truncate();
+            \Solunes\Master\App\AlertConditional::truncate();
+            \Solunes\Master\App\AlertAction::truncate();
+            \Solunes\Master\App\Alert::truncate();
+        }
         // Usuarios
         DB::table('permission_role')->truncate();
         DB::table('role_user')->truncate();  
         \Solunes\Master\App\Role::truncate();
         \Solunes\Master\App\Permission::truncate();
         \App\User::truncate();
-        \Solunes\Master\App\IndicatorValue::truncate();
-        DB::table('indicator_graph_users')->truncate();  
-        \Solunes\Master\App\IndicatorGraph::truncate();
-        DB::table('indicator_alert_users')->truncate();  
-        \Solunes\Master\App\IndicatorAlert::truncate();
-        \Solunes\Master\App\Indicator::truncate();
+        if(config('solunes.indicators')){
+            \Solunes\Master\App\IndicatorUser::truncate();
+            \Solunes\Master\App\Indicator::truncate();
+        }
         // Nodos de Sistema
         \Solunes\Master\App\FieldOptionTranslation::truncate();
         \Solunes\Master\App\FieldOption::truncate();
